chore(apps): create apps folder, move files to proper folders
