+++ client/apps/caption/services/delete-caption.js
@@ -1,4 +1,4 @@
-import {apiService as api } from '../../api/api.service.js';
+import { apiService as api } from "../../services/api.service.js";
 
 const removeCaption = (id) => { 
     api.delete(`delete-caption/${id}`).then(()=> {
