+++ client/apps/meme/components/meme.css
@@ -8,7 +8,7 @@ body {
 
 /* Background Image J.M */
 .bkgnd-container {
-    background: url(../../img/purpgalaxy.jpg);
+    background: url(../../../img/purpgalaxy.jpg);
     color: rgb(0, 0, 0);
     width: 100%;
     height: 100%;
@@ -332,7 +332,7 @@ body {
 }
 
 .modal-img {
-    background-image: url(../../img/ufo\ and\ shaq\ 3.png);
+    background-image: url(../../../img/ufo\ and\ shaq\ 3.png);
     background-size: 100%;
     background-position: right;
     overflow: scroll;
