--- client/caption/components/caption.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-
-    <link rel="stylesheet" href="../components/caption.css">
-    <link rel="stylesheet" href="../../shared/global.css">
-    <!-- CSS Stylesheet -->
-    <script src="../main.caption.js" type="module"></script>
-
-    <!-- Google Fonts Links -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Chau+Philomene+One&family=Roboto&display=swap" rel="stylesheet">
-
-    <title>Captions</title>
-</head>
-<body>
-
-<!-- Background gradient overlay J.M START -->
-<div class="bkgnd-container">
-    <div class="color-overlay"></div>
-  </div>
-  <!--Background gradient overlay J.M END -->
-
-<!-- Table GSA Start -->
-
-<div class="d-flex align-items-center">
-    <nav class="navbar navbar-expand-lg navbar-light navbar-dark text-white fixed-top" style="display: flex; justify-content: space-around">
-        
-        <div class="container-fluid container-nav text-center d-flex">
-            <img class="logo-main" src="../../img/otterhelmet copy.png" alt="" style="width:110px; height: 80px;">
-            
-            <a class="navbar-brand text-white " href="../../index.html" >Meme Stack Generator</a>
-            
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="true" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse  justify-content-end text-center" id="navbarNav">
-                <ul class="navbar-nav topnav-right justify-content-end">
-                    <li class="nav-item">
-                        <a class="nav-link " aria-current="page" href="../../index.html">Home</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link active" href="./caption.html">Captions</a>
-                    </li>
-                </ul>
-            </div>
-        </div>
-        
-    </nav> 
-</div>
-
-<div class=" info-container">
-    <!-- The Modal -->
-    <div id="myModal" class="modal">
-
-        <!-- Modal content -->
-        <div class="container-fluid modal-content">
-        
-                    <!-- <span class="close">&times;</span> -->
-                <div class=" main-container">
-                    <div class="title-container text-center">
-                        <h2>All Captions</h2>
-                    </div>
-                        <div class="main-table-container">
-                            <table id="caption-table">
-                                <tr>
-                                    <th>Top Text</th>
-                                    <th> Bottom Text</th>
-                                    <th> What to do?</th>
-                                
-                                </tr>
-                            </table>
-                        </div>
-                </div>
-                
-            
-        </div>
-    </div>
-
-</div>
-
-
-
-
- 
-    
-    
-<footer class="container-fluid footer-bar">Meme Stack Team &copy; 2022</footer>
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-
-<!-- Table GSA End -->
-</body>
-</html>
