--- client/meme/index.js
@@ -1,8 +1,0 @@
-import * as makeAMeme from '../meme/services/create-meme.js';
-import displayModal from '../meme/services/display-info.js';
-import * as randomizeCaption from '../caption/services/randomize-caption.js'
-
-displayModal();
-
-
-
